requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
langchain>=0.1.0
langchain-openai>=0.1.1
python-dotenv>=1.0.0
lxml>=4.9.3 